fix(authorize): waiting for return of authorization from useCase before executing it
Due to the lack of 'await', the use case was being executed before being validated if the user had
permission to execute it. By default, we always leave the information that the user does not have
access, and we validate it through the 'authorize' method and as we do not wait for this validation,
we return 'not authorized'.